refactor: moved all types from core/src/dom/presentation/types.ts to utils/src/dom/webkit.ts
