+++ packages/core/src/dom/presentation/index.ts
@@ -1,4 +1,3 @@
 export * from './fullscreen';
 export * from './pip';
 export * from './remote-playback';
-export * from './types';
